Simplify object and root query resolvers
