--- test/unit/mirage/resolvers/object.test.ts
@@ -1,112 +1,0 @@
-import { mirageObjectResolver } from '../../../../src/mirage/resolver/object';
-import { generatePackOptions } from '../../../mocks';
-import { GraphQLSchema, buildSchema, GraphQLNonNull, GraphQLString, GraphQLResolveInfo } from 'graphql';
-import { expect } from 'chai';
-import { Model, Server, belongsTo } from 'miragejs';
-
-describe('mirage/resolvers/object', function () {
-  let mirageServer: Server;
-  let schema: GraphQLSchema;
-
-  beforeEach(function () {
-    mirageServer = new Server({
-      models: {
-        user: Model.extend({
-          favoriteMovie: belongsTo('movie'),
-        }),
-        movie: Model.extend({}),
-      },
-    });
-
-    schema = buildSchema(`
-      type User {
-        name: String!
-        favoriteFood: String!
-        favoriteMovie: Movie!
-      }
-
-      type Movie {
-        name: String!
-      }
-    `);
-  });
-
-  it('resolves a simple scalar field from a parent', async function () {
-    const user = mirageServer.create('user', {
-      id: '1',
-      name: 'George',
-    });
-
-    const context = {
-      pack: generatePackOptions({ dependencies: { graphqlSchema: schema } }),
-    };
-
-    const info = {
-      parentType: schema?.getType('User'),
-      fieldName: 'name',
-      returnType: new GraphQLNonNull(GraphQLString),
-    };
-
-    const result = mirageObjectResolver(user, {}, context, info as GraphQLResolveInfo);
-    expect(result).to.equal('George');
-  });
-
-  it('resolves a mirage relationship field from a parent', async function () {
-    const starwars = mirageServer.create('movie', { id: '1', name: 'Star Wars: A New Hope' });
-    const user = mirageServer.create('user', {
-      id: '1',
-      name: 'George',
-      favoriteMovie: starwars,
-    });
-
-    const context = {
-      pack: generatePackOptions({ dependencies: { graphqlSchema: schema } }),
-    };
-
-    const info = {
-      parentType: schema?.getType('User'),
-      fieldName: 'favoriteMovie',
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      returnType: new GraphQLNonNull(schema?.getType('Movie')!),
-    };
-
-    const result = mirageObjectResolver(user, {}, context, info as GraphQLResolveInfo);
-    expect(result.name).to.equal('Star Wars: A New Hope');
-  });
-
-  it('throws an error when a nullish result is returned for a non-null field', function () {
-    const user = mirageServer.create('user', {
-      id: '1',
-    });
-
-    const context = {
-      pack: generatePackOptions({ dependencies: { graphqlSchema: schema } }),
-    };
-
-    const info = {
-      parentType: schema?.getType('User'),
-      fieldName: 'name',
-      returnType: new GraphQLNonNull(GraphQLString),
-    };
-
-    expect(() => mirageObjectResolver(user, {}, context, info as GraphQLResolveInfo)).to.throw(
-      'Failed to resolve field "User.name", got a nullish result for a non-null return type.',
-    );
-  });
-
-  it('throws an error when the parent passed in is not an object', function () {
-    const context = {
-      pack: generatePackOptions({ dependencies: { graphqlSchema: schema } }),
-    };
-
-    const info = {
-      parentType: schema?.getType('User'),
-      fieldName: 'name',
-      returnType: new GraphQLNonNull(GraphQLString),
-    };
-
-    expect(() => mirageObjectResolver('PARENT IS A STRING', {}, context, info as GraphQLResolveInfo)).to.throw(
-      'Expected parent to be an object, got string, when trying to resolve field "name" on type "User"',
-    );
-  });
-});
