+++ test/unit/mirage/type-resolver.test.ts
@@ -1,10 +1,10 @@
 import { Model, Server } from 'miragejs';
-import { generatePackOptions } from '../../../mocks';
+import { generatePackOptions } from '../../mocks';
 import { buildSchema, GraphQLSchema, GraphQLInterfaceType, GraphQLResolveInfo, GraphQLUnionType } from 'graphql';
 import { expect } from 'chai';
-import { mirageTypeResolver } from '../../../../src/mirage';
-import { ResolverInfo, ResolverContext } from '../../../../src/types';
-import { PackOptions } from '../../../../src/pack/types';
+import { mirageTypeResolver } from '../../../src/mirage';
+import { ResolverInfo, ResolverContext } from '../../../src/types';
+import { PackOptions } from '../../../src/pack/types';
 
 function generateContext(packOptions: PackOptions, options?: { useFindInCommon?: boolean }): ResolverContext {
   // important we turn this off for most tests and only
@@ -19,7 +19,7 @@ function generateContext(packOptions: PackOptions, options?: { useFindInCommon?:
   return context;
 }
 
-describe('mirage/resolvers/abstract', function () {
+describe('mirage/type-resolver', function () {
   let mockResolverInfo: ResolverInfo;
   let mirageServer: Server;
